add correct transpilation into the derivative class
