+++ src/squlearn/feature_map/feature_map_derivatives.py
@@ -1,20 +1,25 @@
-from qiskit.circuit import ParameterVector, ParameterExpression
-from qiskit.circuit.parametervector import ParameterVectorElement
-from qiskit.opflow import OperatorBase
-from qiskit.opflow import StateFn, CircuitStateFn, OperatorStateFn
-from qiskit.opflow import SummedOp, ComposedOp, TensoredOp
-from qiskit.opflow import ListOp, PauliOp
-from qiskit.opflow import Zero, One
-from qiskit.compiler import transpile
-from qiskit.quantum_info import Pauli
-from qiskit.opflow import PauliOp
-from qiskit.opflow.gradients import Gradient
-from qiskit.opflow.list_ops.list_op import ListOp as real_ListOp
 import numpy as np
-from typing import Union
+from typing import Union,Set
+
+from qiskit import QuantumCircuit
+from qiskit.circuit import ParameterVector
+from qiskit.circuit.parametervector import ParameterVectorElement
+from qiskit.compiler import transpile  # TODO
 
 from .feature_map_base import FeatureMapBase
 
+from ..util.optree.optree import (
+    OpTreeElementBase,
+    OpTreeLeafCircuit,
+    OpTreeNodeSum,
+    OpTreeNodeList,
+    OpTreeNodeBase,
+)
+from ..util.optree.optree_derivative import simplify_copy, derivative
+from ..util.optree.optree_evaluate import optree_assign_parameters
+from ..util.data_preprocessing import adjust_input
+
+SUPPORTED_GATES = {"x", "y", "z", "h", "rx", "ry", "rz", "p", "cx", "cy", "cz"}
 
 class FeatureMapDerivatives:
     r"""
@@ -85,90 +90,89 @@ class FeatureMapDerivatives:
 
     Args:
         feature_map (FeatureMapBase): Feature map to differentiate
-        opflow_caching (bool): If True, the opflow expressions are cached for faster
+        optree_caching (bool): If True, the optree expressions are cached for faster
                                evaluation. (default: True)
     """
 
     def __init__(
         self,
         feature_map: FeatureMapBase,
-        opflow_caching: bool = True,
+        optree_caching: bool = True,
     ):
         self.feature_map = feature_map
 
         self._x = ParameterVector("x", self.feature_map.num_features)
         self._p = ParameterVector("p", self.feature_map.num_parameters)
 
-        circuit = feature_map.get_circuit(self._x, self._p)
-        self.circuit_opflow = CircuitStateFn(primitive=circuit, coeff=1.0)
-        self.num_qubits = self.circuit_opflow.num_qubits
+        self._circuit = feature_map.get_circuit(self._x, self._p)
 
-        identity_measurement = StateFn(PauliOp(Pauli("I" * self.num_qubits)), is_measurement=True)
-        identity_measurement.replace = True
+        # TODO: remove -> move to parameter-shift implementation!
+        self._instruction_set = list(set(self._circuit.count_ops()))
+        self._circuit = _transpile_to_supported_instructions(self._circuit, SUPPORTED_GATES)
 
-        self.opflow = identity_measurement @ self.circuit_opflow
+        self._optree_start = OpTreeLeafCircuit(self._circuit)
 
-        self.opflow_cache = {}
-        self.opflow_caching = opflow_caching
-        if self.opflow_caching:
-            self.opflow_cache["f"] = self.opflow
+        # self.circuit_optree = CircuitStateFn(primitive=circuit, coeff=1.0)
+        self.num_qubits = self._circuit.num_qubits
+
+        self._optree_cache = {}
+        self._optree_caching = optree_caching
+        if self._optree_caching:
+            self._optree_cache["I"] = OpTreeLeafCircuit(self._circuit)
 
         # get the instruction gates from the initial circuit for transpiling the circuits
         # back to this basis
-        self.instruction_set = set()
-        for instruction in circuit.data:
-            self.instruction_set.add(instruction.operation.name)
-        self.instruction_set = list(self.instruction_set)
 
-    def get_derivative(self, derivative: Union[str, tuple]) -> OperatorBase:
+
+
+    def get_derivative(self, derivative: Union[str, tuple]) -> OpTreeElementBase:
         """Determine the derivative of the feature map circuit.
 
         Args:
             derivative (str or tuple): String or tuple of parameters for specifying the derivation.
 
         Return:
-            Derivative circuit in Qiskit Opflow format.
+            Derivative circuit in OpTree format.
 
         """
         if isinstance(derivative, str):
             if derivative == "I":
-                opflow = self.opflow
+                optree = self._optree_start
             elif derivative == "dx":
-                opflow = self._differentiation_from_tuple((self._x,)).copy()
+                optree = self._differentiation_from_tuple((self._x,)).copy()
             elif derivative == "dxdx":
-                opflow = self._differentiation_from_tuple((self._x, self._x)).copy()
+                optree = self._differentiation_from_tuple((self._x, self._x)).copy()
             elif derivative == "dpdxdx":
-                opflow = self._differentiation_from_tuple((self._p, self._x, self._x)).copy()
+                optree = self._differentiation_from_tuple((self._p, self._x, self._x)).copy()
             elif derivative == "laplace":
                 list_sum = []
                 for xi in self._x:
                     list_sum.append(self._differentiation_from_tuple((xi, xi)).copy())
-                opflow = SummedOp(list_sum)
+                optree = OpTreeNodeSum(list_sum)
             elif derivative == "laplace_dp":
                 list_sum = []
                 for xi in self._x:
                     list_sum.append(self._differentiation_from_tuple((self._p, xi, xi)).copy())
-                opflow = SummedOp(list_sum)
+                optree = OpTreeNodeSum(list_sum)
             elif derivative == "dp":
-                opflow = self._differentiation_from_tuple((self._p,)).copy()
+                optree = self._differentiation_from_tuple((self._p,)).copy()
             elif derivative == "dpdp":
-                opflow = self._differentiation_from_tuple((self._p, self._p)).copy()
+                optree = self._differentiation_from_tuple((self._p, self._p)).copy()
             elif derivative == "dpdx":
-                opflow = self._differentiation_from_tuple((self._p, self._x)).copy()
+                optree = self._differentiation_from_tuple((self._p, self._x)).copy()
             elif derivative == "dxdp":
-                opflow = self._differentiation_from_tuple((self._x, self._p)).copy()
+                optree = self._differentiation_from_tuple((self._x, self._p)).copy()
             else:
                 raise ValueError("Unknown string command:", derivative)
         elif isinstance(derivative, tuple):
-            opflow = self._differentiation_from_tuple(derivative)
+            optree = self._differentiation_from_tuple(derivative)
         else:
             raise TypeError("Input is neither string nor tuple, but:", type(derivative))
 
-        # remove mesurement operator
+        return optree
 
-        return _remove_measure(opflow)
 
-    def get_differentiation_from_tuple(self, diff_tuple: tuple) -> OperatorBase:
+    def get_differentiation_from_tuple(self, diff_tuple: tuple) -> OpTreeElementBase:
         """Returns the derivative of the feature map circuit for a tuple of parameters.
 
         The tuple describes the differentiation with respect to the parameters in the tuple.
@@ -177,11 +181,12 @@ class FeatureMapDerivatives:
             diff_tuple (tuple): Tuple of parameters for specifying the derivation.
 
         Return:
-            Derivative circuit in Qiskit Opflow format.
+            Derivative circuit in OpTree format.
         """
         return self.get_derivative(diff_tuple)
 
-    def get_derivation_from_string(self, input_string: str) -> OperatorBase:
+
+    def get_derivation_from_string(self, input_string: str) -> OpTreeElementBase:
         """Returns the derivative of the feature map circuit for a string abbreviation.
 
         The table for the abbreviations can be found in the documentation of the class.
@@ -190,11 +195,12 @@ class FeatureMapDerivatives:
             input_string (str): String for specifying the derivation.
 
         Return:
-            Derivative circuit in Qiskit Opflow format.
+            Derivative circuit in OpTree format.
         """
         return self.get_derivative(input_string)
 
-    def _differentiation_from_tuple(self, diff_tuple: tuple) -> OperatorBase:
+
+    def _differentiation_from_tuple(self, diff_tuple: tuple) -> OpTreeElementBase:
         """Recursive routine for automatic differentiating the feature map
 
         Variables for the differentiation are supplied by a tuple
@@ -206,26 +212,26 @@ class FeatureMapDerivatives:
                                 determining the derivation
 
         Return:
-            Derivative circuit in Qiskit Opflow format.
+            Derivative circuit in OpTree format.
         """
         # TODO: support tuple of lists of parameter elements
 
         if diff_tuple == ():
-            # Cancel the recursion by returning the opflow of the simply measured feature map
-            return self.opflow.copy()
+            # Cancel the recursion by returning the optree of the simply measured feature map
+            return self._optree_start.copy()
         else:
-            # Check if differentiating tuple is already stored in opflow_cache
-            if self.opflow_caching == True and (diff_tuple,) in self.opflow_cache:
+            # Check if differentiating tuple is already stored in optree_cache
+            if self._optree_caching == True and (diff_tuple,) in self._optree_cache:
                 # If stored -> return
-                return self.opflow_cache[(diff_tuple,)].copy()
+                return self._optree_cache[(diff_tuple,)].copy()
             else:
                 # Recursive differentiation with the most left object
-                circ = _opflow_differentiation(
+                circ = self._optree_differentiation(
                     self._differentiation_from_tuple(diff_tuple[1:]), diff_tuple[0]
                 )
-                # Store result in the opflow_cache
-                if self.opflow_caching == True:
-                    self.opflow_cache[(diff_tuple,)] = circ
+                # Store result in the optree_cache
+                if self._optree_caching == True:
+                    self._optree_cache[(diff_tuple,)] = circ
                 return circ
 
     @property
@@ -233,37 +239,41 @@ class FeatureMapDerivatives:
         """Parameter ParameterVector ``p`` utilized in the feature map circuit."""
         return self._p
 
+
     @property
     def feature_vector(self) -> ParameterVector:
         """Feature ParameterVector ``x`` utilized in the feature map circuit."""
         return self._x
+
 
     @property
     def num_parameters(self) -> int:
         """Number of parameters in the feature map circuit."""
         return len(self._p)
 
+
     @property
     def num_features(self) -> int:
         """Number of features in the feature map circuit."""
         return len(self._x)
 
+
     def assign_parameters(
-        self, opflow: OperatorBase, features: np.ndarray, parameters: np.ndarray
-    ) -> OperatorBase:
+        self, optree: OpTreeElementBase, features: np.ndarray, parameters: np.ndarray
+    ) -> OpTreeElementBase:
         """
         Assigns numerical values to the ParameterVector elements of the feature map circuit.
 
         Args:
-            opflow (OperatorBase): Opflow object to be assigned.
+            optree (OperatorBase): OpTree object to be assigned.
             features (np.ndarray): Numerical values of the feature vector.
             parameters (np.ndarray): Numerical values of the parameter vector.
 
         Return:
-            Opflow object with assigned numerical values.
+            OpTree object with assigned numerical values.
         """
 
-        if opflow is None:
+        if optree is None:
             return None
 
         todo_list = []  # list for the variables
@@ -272,12 +282,12 @@ class FeatureMapDerivatives:
 
         # check shape of the x and adjust to [[]] form if necessary
         if features is not None:
-            xx, multi_x = _adjust_input(features, len(self._x))
+            xx, multi_x = adjust_input(features, len(self._x))
             todo_list.append(xx)
             param_list.append(self._x)
             multi_list.append(multi_x)
         if parameters is not None:
-            pp, multi_p = _adjust_input(parameters, len(self._p))
+            pp, multi_p = adjust_input(parameters, len(self._p))
             todo_list.append(pp)
             param_list.append(self._p)
             multi_list.append(multi_p)
@@ -295,290 +305,109 @@ class FeatureMapDerivatives:
                         rec_assign(dic.copy(), todo_list[1:], param_list[1:], multi_list[1:])
                     )
                 else:
-                    return_list.append(opflow.assign_parameters(dic))
+                    return_list.append(optree_assign_parameters(optree, dic))
 
             if multi_list[0]:
-                return ListOp(return_list)
+                return OpTreeNodeList(return_list)
             else:
                 return return_list[0]
 
         return rec_assign({}, todo_list, param_list, multi_list)
 
 
-def _adjust_input(x, x_length) -> tuple:
-    """Adjust the input to the form [[]] if necessary.
+    def _optree_differentiation(
+        self,
+        optree: OpTreeElementBase,
+        parameters: Union[list, tuple, ParameterVectorElement, ParameterVector],
+    ) -> OpTreeElementBase:
+        """
+        Routine for the automatic differentiation based on qiskit routines
 
-    Args:
-        x (np.ndarray): Input array.
-        x_length (int): Dimension of the input array, e.g. feature dimension.
+        Args:
+            optree : Input OpTree expression
+            params (list | ParameterVector): variables which are used in the
+                                                differentiation (have to be the same type )
+        Returns:
+            OpTree structure of the differentiated input optree
+        """
 
-    Return:
-        Adjusted input array and a boolean flag for multiple inputs.
-    """
+        # Make list if input is not a list
+        if isinstance(parameters, ParameterVectorElement):
+            parameters = [parameters]
+        if isinstance(parameters, tuple):
+            parameters = list(parameters)
 
-    # check shape of the x and adjust to [[]] form if necessary
-    multiple_inputs = False
-    error = False
-    shape = np.shape(x)
-    if shape == () and x_length == 1:
-        # Single floating point number
-        xx = np.array([[x]])
-    elif len(shape) == 1:
-        if x_length == 1:
-            xx = np.array([np.array([xx]) for xx in x])
-            multiple_inputs = True
+        # Call the automatic differentiation routine
+        # Separate routine for for one dim. or multi dim. variables
+        if len(parameters) == 1:
+            return _optree_transpile_back(simplify_copy(
+                derivative(optree, parameters).children[0]
+            ),self._instruction_set)  # TODO:backtranspile maybe in derivative
         else:
-            # We have a single multi dimensional x (e.g. parameter vector)
-            if len(x) == x_length:
-                xx = np.array([x])
-            else:
-                error = True
-    elif len(shape) == 2:
-        if shape[1] == x_length:
-            xx = x
-            multiple_inputs = True
+            # If multiple variables are differentiated -> results are returned in array
+
+            # Check if the same variables are the same type
+            params_name = parameters[0].name.split("[", 1)[0]
+            for p in parameters:
+                if p.name.split("[", 1)[0] != params_name:
+                    raise TypeError("Differentiable variables are not the same type.")
+
+            return _optree_transpile_back(simplify_copy(
+                derivative(optree, parameters)
+            ),self._instruction_set)  # TODO:backtranspile maybe in derivative
+
+
+def _optree_transpile_back(
+    optree_element: Union[OpTreeNodeBase, OpTreeLeafCircuit, QuantumCircuit], instruction_set
+) -> Union[OpTreeNodeBase, OpTreeLeafCircuit, QuantumCircuit]:
+    if isinstance(optree_element, OpTreeNodeBase):
+        # Recursive call for all children
+        children_list = [
+            _optree_transpile_back(child, instruction_set) for child in optree_element.children
+        ]
+        if isinstance(optree_element, OpTreeNodeSum):
+            return OpTreeNodeSum(children_list, optree_element.factor, optree_element.operation)
+        elif isinstance(optree_element, OpTreeNodeList):
+            return OpTreeNodeList(children_list, optree_element.factor, optree_element.operation)
         else:
-            error = True
+            raise ValueError("element must be a CircuitTreeSum or a CircuitTreeList")
+    elif isinstance(optree_element, (OpTreeLeafCircuit, QuantumCircuit)):
+        circuit = optree_element
+        if isinstance(optree_element, OpTreeLeafCircuit):
+            circuit = optree_element.circuit
+
+        # Transpile back to the given instruction set
+        transpiled_circ = transpile(
+            circuit,
+            basis_gates=instruction_set,
+            optimization_level=1,  # 1 for reducing number of gates
+            layout_method="trivial",
+        )
+
+        if isinstance(optree_element, OpTreeLeafCircuit):
+            return OpTreeLeafCircuit(transpiled_circ)
+
+        return transpiled_circ
+
     else:
-        error = True
+        raise ValueError("Unsupported type in _optree_transpile_back:", type(optree_element))
 
-    if error:
-        raise ValueError("Wrong format of an input variable.")
-
-    return xx, multiple_inputs
-
-
-def _remove_measure(operator: OperatorBase) -> OperatorBase:
-    """Replace all measurements from the inputted Opflow expression
+def _transpile_to_supported_instructions(
+    circuit: QuantumCircuit, supported_gates: Set[str]
+) -> QuantumCircuit:
+    """ Helper function for transpiling a circuit to a supported instruction set.
 
     Args:
-        operator (OperatorBase): Opflow expression from which the measurements are removed
+        circuit (QuantumCircuit): Circuit to transpile.
+        supported_gates (Set[str]): Set of supported gates.
 
-    Return:
-        New opflow expression with replaced measurement operator
-    """
-    # We reached a ComposedOp term -> replace the measurement
-    if isinstance(operator, ComposedOp):
-        # Remove the measurement from the children
-        for i in range(len(operator.oplist)):
-            if isinstance(operator.oplist[i], CircuitStateFn):
-                return_op = operator.oplist[i]
-                return_op._coeff = operator.coeff
-                return return_op
-
-        return ~Zero @ One
-
-    # We reached a CircuitStateFn term -> apply the measurement to the wave function
-    elif isinstance(operator, CircuitStateFn):
-        return operator
-
-    # We reached a list of opflow terms, recursive call of
-    # the function to all children in the list
-    elif isinstance(operator, ListOp):
-        op_list = [_remove_measure(op) for op in operator.oplist]
-
-        if isinstance(operator, SummedOp):
-            return SummedOp(
-                oplist=[op for op in op_list if op != ~Zero @ One], coeff=operator.coeff
-            )
-        elif isinstance(operator, TensoredOp):
-            return TensoredOp(
-                oplist=[op for op in op_list if op != ~Zero @ One], coeff=operator.coeff
-            )
-        elif isinstance(operator, real_ListOp):
-            return ListOp(
-                oplist=[op for op in op_list if op != ~Zero @ One],
-                coeff=operator.coeff,
-                combo_fn=operator.combo_fn,
-            )
-        else:
-            raise ValueError("Unknown ListOp type in _remove_measure:", type(operator))
-    else:
-        raise ValueError("Unknown type in _remove_measure:", type(operator))
-
-
-def measure_feature_map_derivative(
-    operator: OperatorBase, measurement: OperatorBase
-) -> OperatorBase:
-    """
-    Applied the measurement operator to each circuit in the given opflow structure.
-
-    Args:
-        operator (OperatorBase): Opflow expression that is measured
-        measurement (OperatorBase): Measurement operator
-
-    Return:
-        New opflow expression with replaced measurement operator
-    """
-    # We reached a ComposedOp term -> replace the measurement
-    if isinstance(operator, ComposedOp):
-        for i in range(len(operator.oplist)):
-            if isinstance(operator.oplist[i], OperatorStateFn):
-                if hasattr(operator.oplist[i], "replace"):
-                    if operator.oplist[i].replace:
-                        operator.oplist[i] = measurement
-                else:
-                    operator.oplist[i] = measurement
-        return operator
-
-    # We reached a CircuitStateFn term -> apply the measurement to the wave function
-    elif isinstance(operator, CircuitStateFn):
-        coeff = operator._coeff
-        operator._coeff = 1.0
-        measured_op = measurement @ operator
-        measured_op._coeff = coeff
-        return measured_op
-
-    # We reached a list of opflow terms, recursive call of
-    # the function to all children in the list
-    elif isinstance(operator, ListOp):
-        op_list = [measure_feature_map_derivative(op, measurement) for op in operator.oplist]
-
-        if isinstance(operator, SummedOp):
-            return SummedOp(
-                oplist=[op for op in op_list if op != ~Zero @ One], coeff=operator.coeff
-            )
-        elif isinstance(operator, TensoredOp):
-            return TensoredOp(
-                oplist=[op for op in op_list if op != ~Zero @ One], coeff=operator.coeff
-            )
-        elif isinstance(operator, real_ListOp):
-            return ListOp(
-                oplist=[op for op in op_list if op != ~Zero @ One],
-                coeff=operator.coeff,
-                combo_fn=operator.combo_fn,
-            )
-        else:
-            raise ValueError("Unknown ListOp type in measure_derivative:", type(operator))
-    else:
-        raise ValueError("Unknown type in measure_derivative:", type(operator))
-
-
-def _opflow_differentiation(
-    opflow: OperatorBase, parameters: Union[list, tuple, ParameterVectorElement, ParameterVector]
-):
-    """
-    Routine for the automatic differentiation based on qiskit routines
-
-    Args:
-        opflow : Input opflow expression
-        params (list | ParameterVector): variables which are used in the
-                                            differentiation (have to be the same type )
     Returns:
-        opflow structure of the differentiated input opflow
+        Circuit which is transpiled to the supported instruction set.
     """
 
-    # Make list if input is not a list
-    if isinstance(parameters, ParameterVectorElement):
-        parameters = [parameters]
-    if isinstance(parameters, tuple):
-        parameters = list(parameters)
-
-    # Call the automatic differentiation routine
-    # Separate routine for for one dim. or multi dim. variables
-    if len(parameters) == 1:
-        # Automatic qiskit differentiation
-        # ParameterExpression(0) is removed by clean_opflow_circ
-        # Gates are transpiled in Gradient(), this can yield different set of gates
-        # than supported by the QC hardware, this is also fixed by clean_opflow_circ
-        return _clean_opflow_circ(Gradient().convert(operator=opflow, params=parameters)[0])
-    else:
-        # If multiple variables are differentiated -> results are returned in array
-
-        # Check if the same variables are the same type
-        params_name = parameters[0].name.split("[", 1)[0]
-        for p in parameters:
-            if p.name.split("[", 1)[0] != params_name:
-                raise TypeError("Differentiable variables are not the same type.")
-
-        # Automatic qiskit differentiation
-        # ParameterExpression(0) is removed by clean_opflow_circ
-        # Gates are transpiled in Gradient(), this can yield different set of gates
-        # than supported by the QC hardware, this is also fixed by clean_opflow_circ
-        return _clean_opflow_circ(Gradient().get_gradient(operator=opflow, params=parameters))
-
-
-def _clean_opflow_circ(operator: OperatorBase, instruction_set: list[str] = None) -> OperatorBase:
-    """
-    Function for cleaning the opflow structure.
-
-    - Removes all ParameterExpression(0) contributions
-    - Transpiling back to the given instruction set
-
-    Args:
-        operator (OperatorBase): input opflow expression
-        instruction_set (list): list of gates which are supported by the QC hardware
-
-    Return:
-        The opflow expression without the zero contributions and transpiled circuits
-    """
-
-    # We reached the Composed object or the wave function
-    if isinstance(operator, CircuitStateFn):
-        # If coefficient is zero replace by Zero @ One which is sorted out later
-        ZeroParameterExpr = False
-        if isinstance(operator.coeff, ParameterExpression):
-            # Check for ParameterExpression(0)
-            ZeroParameterExpr = operator.coeff.sympify() == 0
-        if operator.coeff != 0.0 and operator.coeff != 0 and ZeroParameterExpr == False:
-            if instruction_set is None:
-                return operator
-            else:
-                # Transpile back to the original basis set
-                transpiled_circ = transpile(
-                    operator.primitive,
-                    basis_gates=instruction_set,
-                    optimization_level=1,  # 1 for reducing number of gates
-                    layout_method="trivial",
-                )
-
-                transpiled_circ = operator.primitive
-                return CircuitStateFn(
-                    primitive=transpiled_circ,
-                    coeff=operator.coeff,
-                    is_measurement=operator.is_measurement,
-                )
-        else:
-            return ~Zero @ One
-    elif isinstance(operator, OperatorStateFn):
-        # Measurement operator -> do nothing
-        return operator
-    elif isinstance(operator, ComposedOp):
-        # If coefficient is zero replace by Zero @ One which is sorted out later
-        ZeroParameterExpr = False
-        if isinstance(operator.coeff, ParameterExpression):
-            # Check for ParameterExpression(0)
-            ZeroParameterExpr = operator.coeff.sympify() == 0
-        if operator.coeff != 0.0 and operator.coeff != 0 and ZeroParameterExpr == False:
-            # None-zero coefficient found, iterate through composed operators
-            op_list = [_clean_opflow_circ(op, instruction_set) for op in operator.oplist]
-            if ~Zero @ One in op_list:
-                return ~Zero @ One
-            else:
-                return ComposedOp(op_list, coeff=operator.coeff)
-        else:
-            return ~Zero @ One
-
-    elif isinstance(operator, ListOp):
-        # List object reached, recursive call of the function
-        op_list = [_clean_opflow_circ(op, instruction_set) for op in operator.oplist]
-
-        # Sort out the Zero @ One terms
-        if isinstance(operator, SummedOp):
-            return SummedOp(
-                oplist=[op for op in op_list if op != ~Zero @ One], coeff=operator.coeff
-            )
-        elif isinstance(operator, TensoredOp):
-            return TensoredOp(
-                oplist=[op for op in op_list if op != ~Zero @ One], coeff=operator.coeff
-            )
-        elif isinstance(operator, real_ListOp):
-            return ListOp(
-                oplist=[op for op in op_list if op != ~Zero @ One],
-                coeff=operator.coeff,
-                combo_fn=operator.combo_fn,
-            )
-        else:
-            raise ValueError("Unknown Type in ListOp")
-    else:
-        raise ValueError("Unknown Type in OperatorBase:", type(operator))
+    unique_ops = set(circuit.count_ops())
+    if not unique_ops.issubset(supported_gates):
+        circuit = transpile(
+            circuit, basis_gates=list(supported_gates), optimization_level=0, layout_method="trivial"
+        )
+    return circuit
